move common traits
