+++ OpenRA.Mods.Cnc/OpenRA.Mods.Cnc.csproj
@@ -75,7 +75,6 @@
     <Compile Include="Traits\Render\RenderGunboat.cs" />
     <Compile Include="Traits\Render\WithCargo.cs" />
     <Compile Include="Traits\Render\WithDeliveryAnimation.cs" />
-    <Compile Include="Traits\Render\WithFire.cs" />
     <Compile Include="Traits\Render\WithRoof.cs" />
     <Compile Include="Traits\SpawnViceroid.cs" />
     <Compile Include="Traits\SupportPowers\IonCannonPower.cs" />
