+++ OpenRA.Mods.Common/Activities/DonateSupplies.cs
@@ -8,11 +8,10 @@
  */
 #endregion
 
-using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Effects;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Activities
+namespace OpenRA.Mods.Common.Activities
 {
 	class DonateSupplies : Enter
 	{
