+++ OpenRA.Mods.Common/Traits/Attack/AttackGarrisoned.cs
@@ -13,10 +13,9 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Activities;
 using OpenRA.Graphics;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Traits
+namespace OpenRA.Mods.Common.Traits
 {
 	public class FirePort
 	{
