+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -65,54 +65,47 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Traits\AcceptsSupplies.cs" />
-    <Compile Include="Activities\DonateSupplies.cs" />
     <Compile Include="Activities\Infiltrate.cs" />
     <Compile Include="Activities\LayMines.cs" />
     <Compile Include="Activities\Leap.cs" />
     <Compile Include="Activities\Teleport.cs" />
+    <Compile Include="Effects\GpsDot.cs" />
     <Compile Include="Effects\GpsSatellite.cs" />
     <Compile Include="Effects\SatelliteLaunch.cs" />
     <Compile Include="Effects\TeslaZap.cs" />
-    <Compile Include="Traits\Render\RenderUnitReload.cs" />
     <Compile Include="Graphics\TeslaZapRenderable.cs" />
-    <Compile Include="Traits\EjectOnDeath.cs" />
     <Compile Include="Traits\Attack\AttackLeap.cs" />
-    <Compile Include="Traits\PaletteEffects\ChronoshiftPaletteEffect.cs" />
-    <Compile Include="Traits\Chronoshiftable.cs" />
-    <Compile Include="Traits\DemoTruck.cs" />
-    <Compile Include="Effects\GpsDot.cs" />
-    <Compile Include="Traits\MadTank.cs" />
-    <Compile Include="Traits\Mine.cs" />
-    <Compile Include="Traits\Minelayer.cs" />
-    <Compile Include="Traits\PortableChrono.cs" />
-    <Compile Include="Traits\Render\RenderJammerCircle.cs" />
-    <Compile Include="Traits\Render\RenderDisguise.cs" />
-    <Compile Include="Traits\Render\RenderLandingCraft.cs" />
-    <Compile Include="Traits\Disguise.cs" />
-    <Compile Include="Traits\SupplyTruck.cs" />
-    <Compile Include="Traits\SupportPowers\ChronoshiftPower.cs" />
-    <Compile Include="Traits\SupportPowers\GpsPower.cs" />
-    <Compile Include="Traits\SupportPowers\ParatroopersPower.cs" />
     <Compile Include="Traits\Buildings\Fake.cs" />
+    <Compile Include="Traits\Buildings\ClonesProducedUnits.cs" />
+    <Compile Include="Traits\Chronoshiftable.cs" />
+    <Compile Include="Traits\Cloneable.cs" />
+    <Compile Include="Traits\DemoTruck.cs" />
+    <Compile Include="Traits\Disguise.cs" />
     <Compile Include="Traits\HarvesterHuskModifier.cs" />
-    <Compile Include="Traits\LeavesHusk.cs" />
-    <Compile Include="Traits\TargetableSubmarine.cs" />
-    <Compile Include="Traits\TransformOnPassenger.cs" />
-    <Compile Include="Traits\Render\RenderShroudCircle.cs" />
     <Compile Include="Traits\Infiltration\InfiltrateForCash.cs" />
     <Compile Include="Traits\Infiltration\InfiltrateForExploration.cs" />
     <Compile Include="Traits\Infiltration\InfiltrateForPowerOutage.cs" />
     <Compile Include="Traits\Infiltration\InfiltrateForSupportPower.cs" />
     <Compile Include="Traits\Infiltration\Infiltrates.cs" />
-    <Compile Include="CombatDebugOverlay.cs" />
-    <Compile Include="Traits\Attack\AttackGarrisoned.cs" />
-    <Compile Include="Scripting\Properties\ChronosphereProperties.cs" />
-    <Compile Include="Traits\Buildings\ClonesProducedUnits.cs" />
-    <Compile Include="Traits\Cloneable.cs" />
-    <Compile Include="Scripting\Properties\HarvesterProperties.cs" />
-    <Compile Include="Scripting\Properties\TransportProperties.cs" />
     <Compile Include="Traits\InvulnerabilityUpgrade.cs" />
+    <Compile Include="Traits\MadTank.cs" />
+    <Compile Include="Traits\Mine.cs" />
+    <Compile Include="Traits\Minelayer.cs" />
+    <Compile Include="Traits\PaletteEffects\ChronoshiftPaletteEffect.cs" />
+    <Compile Include="Traits\PortableChrono.cs" />
+    <Compile Include="Traits\Render\RenderJammerCircle.cs" />
+    <Compile Include="Traits\Render\RenderDisguise.cs" />
+    <Compile Include="Traits\Render\RenderLandingCraft.cs" />
+    <Compile Include="Traits\Render\RenderShroudCircle.cs" />
+    <Compile Include="Traits\Render\RenderUnitReload.cs" />
+    <Compile Include="Traits\SupportPowers\ChronoshiftPower.cs" />
+    <Compile Include="Traits\SupportPowers\GpsPower.cs" />
+    <Compile Include="Traits\SupportPowers\ParatroopersPower.cs" />
+    <Compile Include="Traits\LeavesHusk.cs" />
+    <Compile Include="Traits\TargetableSubmarine.cs" />
+    <Compile Include="Scripting\Properties\ChronosphereProperties.cs" />
+    <Compile Include="Scripting\Properties\ParadropProperties.cs" />
+    <Compile Include="Scripting\Properties\ParatroopersProperties.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\OpenRA.Game\OpenRA.Game.csproj">
