+++ mods/cnc/mod.yaml
@@ -86,7 +86,6 @@ Assemblies:
 	./mods/common/OpenRA.Mods.Common.dll
 	./mods/ra/OpenRA.Mods.RA.dll
 	./mods/cnc/OpenRA.Mods.Cnc.dll
-	./mods/d2k/OpenRA.Mods.D2k.dll
 
 ChromeLayout:
 	./mods/cnc/chrome/install.yaml
